+++ test/unit-complete/main.tf
@@ -1,34 +1,3 @@
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# COMPLETE FEATURES UNIT TEST
-# This module tests a complete set of most/all non-exclusive features
-# The purpose is to activate everything the module offers, but trying to keep execution time and costs minimal.
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-variable "gcp_region" {
-  type        = string
-  description = "(Required) The gcp region in which all resources will be created."
-}
-
-variable "gcp_project" {
-  type        = string
-  description = "(Required) The ID of the project in which the resource belongs."
-}
-
-terraform {
-  required_providers {
-    google-beta = {
-      source  = "hashicorp/google-beta"
-      version = "~> 4.0"
-    }
-  }
-}
-
-provider "google-beta" {
-  region  = var.gcp_region
-  project = var.gcp_project
-}
-
-# DO NOT RENAME MODULE NAME
 module "test" {
   source = "../.."
 
@@ -37,7 +6,7 @@ module "test" {
   # add all required arguments
   repository_id = "unit-complete"
   format        = "NPM"
-  location      = var.gcp_region
+  location      = "europe-west3"
 
   # add all optional arguments that create additional resources
   iam = [
@@ -67,8 +36,3 @@ module "test" {
   module_depends_on = ["nothing"]
 }
 
-# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
-# output "all" {
-#   description = "All outputs of the module."
-#   value       = module.test
-# }