+++ variables.tf
@@ -8,15 +8,17 @@ variable "repository_id" {
   type        = string
 }
 
-variable "location" {
-  description = "(Required) The name of the location this repository is located in."
-  type        = string
-}
-
 # ---------------------------------------------------------------------------------------------------------------------
 # OPTIONAL VARIABLES
 # These variables have defaults, but may be overridden.
 # ---------------------------------------------------------------------------------------------------------------------
+
+variable "location" {
+  description = "(Optional) The name of the location this repository is located in. If it is not provided, the provider location is used."
+  type        = string
+  default     = null
+}
+
 
 variable "format" {
   description = "(Optional) The format of packages that are stored in the repository. You can only create alpha formats if you are a member of the alpha user group. Possible values are `DOCKER`, `MAVEN`, `NPM`, `PYTHON`, `APT` (alpha), `YUM` (alpha)"
@@ -32,8 +34,8 @@ variable "description" {
 
 variable "labels" {
   description = "(Optional) Labels with user-defined metadata. This field may contain up to 64 entries. Label keys and values may be no longer than 63 characters. Label keys must begin with a lowercase letter and may only contain lowercase letters, numeric characters, underscores, and dashes."
-  type        = any
-  default     = null
+  type        = map(string)
+  default     = {}
 }
 
 variable "kms_key_name" {